break out individual pages / elements in POM and BOM
